(chore) add class functionality
